pyproject.toml
[project]
name = "finalproject-menshikova-daria-dpo-nod"
version = "0.1.0"
description = ""
authors = [
    {name = "Daria Menshikova",email = "dari.menshikova@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["prompt (>=0.4.1,<0.5.0)", "tomli (>=2.3.0,<3.0.0)", "python-dotenv (>=1.2.1,<2.0.0)", "requests (>=2.31.0,<3.0.0)"]

[tool.poetry]
packages = [
    {include = "valutatrade_hub"}
]

[tool.valutatrade]
data_directory = "data/"
rates_ttl_seconds = 300
default_base_currency = "USD"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
ruff = "^0.13.2"

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["venv", ".venv", "dist", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.poetry.scripts]
project = "valutatrade_hub.cli.main:main"

